backend/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.8.0
Pillow==9.0
requests==2.27.1
flake8==4.0.1
tox==3.24.5
django-storages==1.12.3

psycopg2-binary==2.9.3 # binary can reduce performance ?
gunicorn==20.1.0

Celery==5.2.6
redis==4.2.2

# websocket
channels==4.0.0
daphne==4.0.0
asgiref==3.6.0
channels-redis==4.0.0